Prepare a bit for unit testing
