+++ Tests/gtest/samples/playground.h
@@ -1,0 +1,7 @@
+#pragma once
+
+// Returns n! (the factorial of n).  For negative n, n! is defined to be 1.
+int Factorial(int n);
+
+// Returns true iff n is a prime number.
+bool IsPrime(int n);
